first working version

+++ server/src/controllers/users.ts
@@ -9,6 +9,9 @@ import type {
   DeleteUserParams,
   GetAllUsersQuery,
   LoginBody,
+  UpdateUserBody,
+  UpdateUserParams,
+  UpdateUserSettingsBody,
 } from "../schemas/user.schemas";
 
 export const createDevAdmin = async (
@@ -19,12 +22,45 @@ export const createDevAdmin = async (
     const saltRounds = 10;
     const hashedPassword = await bcrypt.hash("admin", saltRounds);
     const newAdmin = await prisma.user.create({
-      data: { username: "admin", password: hashedPassword, role: "ADMIN" },
+      data: {
+        username: "admin",
+        password: hashedPassword,
+        role: "ADMIN",
+        settings: {},
+      },
     });
     return reply.code(201).send(newAdmin);
   } catch (error) {
     request.log.error(error);
     return reply.code(500).send({ error: "couldn't create admin" });
+  }
+};
+
+export const updateUserRole = async (
+  request: FastifyRequest<{
+    Body: UpdateUserBody;
+    Params: UpdateUserParams;
+  }>,
+  reply: FastifyReply,
+) => {
+  const { role } = request.body;
+  const { id: userId } = request.params;
+
+  try {
+    const updatedUser = await prisma.user.update({
+      where: {
+        id: userId,
+      },
+      data: {
+        role: role,
+      },
+    });
+    return reply.code(200).send(updatedUser);
+  } catch (error) {
+    request.log.error(error);
+    return reply
+      .code(500)
+      .send({ error: `couldn't update user with id: ${userId}` });
   }
 };
 
@@ -91,7 +127,7 @@ export const registerUser = async (
 
   try {
     const newUser = await prisma.user.create({
-      data: { username, password: hashedPassword },
+      data: { username, password: hashedPassword, settings: {} },
     });
     const { password: _, ...userWithoutPassword } = newUser;
     return reply.code(201).send(userWithoutPassword);
@@ -107,6 +143,15 @@ export const deleteUser = async (
 ) => {
   const { id } = request.params;
   try {
+    const user = await prisma.user.findUnique({ where: { id } });
+    if (user === null) {
+      return reply.code(404).send({ message: "this user does not exist" });
+    }
+    if (user.role === "ADMIN") {
+      return reply
+        .code(403)
+        .send({ message: "ADMIN users can not be deleted" });
+    }
     await prisma.user.delete({ where: { id } });
     return reply.code(204).send({ message: "deleted user successfully" });
   } catch (error) {
@@ -164,4 +209,44 @@ export const logout = async (request: FastifyRequest, reply: FastifyReply) => {
 
 export const getMe = async (request: FastifyRequest, reply: FastifyReply) => {
   return reply.code(200).send({ user: request.user });
+};
+
+export const updateUserSettings = async (
+  request: FastifyRequest<{ Body: UpdateUserSettingsBody }>,
+  reply: FastifyReply,
+) => {
+  const { uiLang, theme } = request.body;
+  const user = request.user;
+
+  try {
+    const updatedUserSettings = await prisma.userSettings.update({
+      where: { userId: user.id },
+      data: {
+        uiLang: uiLang,
+        theme: theme,
+      },
+    });
+
+    return reply.code(200).send(updatedUserSettings);
+  } catch (error) {
+    request.log.error(error);
+    return reply.code(500).send({ error: "failed to update user settings" });
+  }
+};
+
+export const getUserSettings = async (
+  request: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  const user = request.user;
+
+  try {
+    const userSettings = await prisma.userSettings.findFirst({
+      where: { userId: user.id },
+    });
+    return reply.code(200).send(userSettings);
+  } catch (error) {
+    request.log.error(error);
+    return reply.code(500).send({ error: "failed to fetch user settings" });
+  }
 };
